+++ Sources/App/Modules/Frontend/Controllers/FrontendController.swift
@@ -12,10 +12,7 @@ struct FrontendController {
             )))
         }
         
-        // Check expiration and delete expired tokens
         guard token.expiresAt > .now else {
-            // Delete expired email tokens (they are useless once expired)
-            try await req.repositories.emailTokens.delete(id: token.requireID())
             return req.templates.renderHtml(OutcomeMessageTemplate(.init(
                 text: "Token expired, please request verification again.",
                 title: "Token expired"
