--- Tests/AppTests/Framework/Helpers/XCTAssertResponseError.swift
@@ -1,12 +1,0 @@
-@testable import App
-import VaporTesting
-import XCTest
-import Testing
-
-func XCTAssertResponseError(_ res: TestingHTTPResponse, _ error: AppError, file: StaticString = #filePath, line: UInt = #line) {
-    XCTAssertEqual(res.status, error.status, file: file, line: line)
-    expectContent(ErrorResponse.self, res) { errorContent in
-        XCTAssertEqual(errorContent.errorIdentifier, error.identifier, file: file, line: line)
-        XCTAssertEqual(errorContent.reason, error.reason, file: file, line: line)
-    }
-}
