--- main 2.xctestplan
@@ -1,18 +1,0 @@
-{
-  "configurations" : [
-    {
-      "id" : "F42A432F-E416-46F9-95DE-1CBFBC629C0F",
-      "name" : "Configuration 1",
-      "options" : {
-
-      }
-    }
-  ],
-  "defaultOptions" : {
-    "testTimeoutsEnabled" : true
-  },
-  "testTargets" : [
-
-  ],
-  "version" : 1
-}
